requirements.txt
google-genai
angchain
langchain[google-genai]"
langchain_google_genai